ATMega328p grid assembly
With this 'atm328p_grid.asm' file you have the starter kit for a 10ms repeated and auto reload assembly program. On this base you have the right grid to make all your assembly projects for the ATMega328p micro controller. The assembled atm328p:grid.hex is the to program file for the micro-controller.